feat(routing): routing now begins at login screen
any user name or password is accepted, even blank
