+++ navigation/AppNavigator.js
@@ -1,10 +1,16 @@
-import React from 'react';
 import { createAppContainer, createSwitchNavigator } from 'react-navigation';
 
 import MainTabNavigator from './MainTabNavigator';
+import AuthLoadingScreen from '../screens/AuthLoadingScreen';
+import LoginScreen from '../screens/AuthLoginScreen';
 
-export default createAppContainer(createSwitchNavigator({
-  // You could add another route here for authentication.
-  // Read more at https://reactnavigation.org/docs/en/auth-flow.html
-  Main: MainTabNavigator,
-}));
+export default createAppContainer(createSwitchNavigator(
+  {
+    AuthLoading: AuthLoadingScreen,
+    App: MainTabNavigator,
+    Auth: LoginScreen,
+  },
+  {
+    initialRouteName: 'AuthLoading',
+  }
+));
