+++ navigation/MainTabNavigator.js
@@ -7,7 +7,6 @@ import HomeScreen from '../screens/HomeScreen';
 import FriendsScreen from '../screens/FriendsScreen';
 import SettingsScreen from '../screens/SettingsScreen';
 import ProfileScreen from '../screens/ProfileScreen';
-import LoginScreen from '../screens/LoginScreen';
 
 const HomeStack = createStackNavigator({
   Home: HomeScreen,
@@ -77,24 +76,9 @@ ProfileStack.navigationOptions = {
   ),
 };
 
-const LoginStack = createStackNavigator({
-  Login: LoginScreen,
-});
-
-LoginStack.navigationOptions = {
-  tabBarLabel: "Login",
-  tabBarIcon: ({ focused }) => (
-    <TabBarIcon
-      focused={focused}
-      name={Platform.OS === "ios" ? "ios-log-in" : "md-log-in"}
-    />
-  ),
-};
-
 export default createBottomTabNavigator({
   HomeStack,
   FriendsStack,
   ProfileStack,
-  LoginStack,
   SettingsStack
 });
