+++ screens/ProfileScreen.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { ScrollView, StyleSheet, Text, View, Image, FlatList } from 'react-native';
-import { List, ListItem, Avatar, Icon, Button   } from "react-native-elements";
+import { List, ListItem, Avatar, Icon, Button } from "react-native-elements";
 import theme from '../constants/Colors';
 
 function getRandomActivity(randomInt) {
@@ -95,7 +95,7 @@ export default class ProfileScreen extends React.Component {
           <Button
             loading={this.state.loading}
             buttonStyle={{
-              backgroundColor: "rgba(187,0,0,1)",
+              backgroundColor: theme.primaryColor,
               width: 300,
               height: 45,
               borderColor: 'transparent',
